+++ activemq-commons/src/main/java/org/apache/activemq/logs/ActiveMQUtilBundle.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.activemq.utils;
+package org.apache.activemq.logs;
 
 
 import org.apache.activemq.api.core.ActiveMQIllegalStateException;
