+++ activemq-commons/src/main/java/org/apache/activemq/utils/PasswordMaskingUtil.java
@@ -23,6 +23,7 @@ import java.util.Map;
 
 import org.apache.activemq.api.core.ActiveMQException;
 import org.apache.activemq.api.core.ActiveMQExceptionType;
+import org.apache.activemq.logs.ActiveMQUtilBundle;
 
 /**
  * A PasswordMarkingUtil
