+++ activemq-server/src/test/resources/logging.properties
@@ -45,7 +45,7 @@ handler.FILE.formatter=PATTERN
 
 
 # Console handler configuration
-handler.TEST=org.apache.activemq.utils.AssertionLoggerHandler
+handler.TEST=org.apache.activemq.logs.AssertionLoggerHandler
 handler.TEST.level=TRACE
 handler.TEST.formatter=PATTERN
